Added gnplib.ui.helper namespace and the drawButton function.
This namespace contains helper functions for the syntactically clean drawing of ui elements inside of the gnplib.ui namespace.

The drawButton function cleanly wraps up the easeljs drawing of a roundRect with a given color, while leaving the door open for future appended draw calls.

Also cleaned up contents of createButtonSimple to make use of drawButton.